+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Made API endpoint configurable
+- Added [agent command](./README.md#agent-command)
 
 ## [0.1.0] - 2024-12-09